+++ src/main/sensors/Thermometer.h
@@ -7,13 +7,13 @@
 
 class Thermometer {
 public:
-	int potVal = 0;
+	double potVal = 0;
 	int potPin = 35;
 	LED* led;
 
 	Thermometer();
 	Thermometer(int potPin, LED *led);
-	TemperatureStatus calculateStatus(int potVal, LED *led);
+	TemperatureStatus calculateStatus(double potVal, LED *led);
 	String getTempStatus(TemperatureStatus currentStatus);
 };
 #endif
